+++ source/SAMRAI/pdat/OuternodeData.h
@@ -422,7 +422,7 @@ public:
     *
     * @pre dynamic_cast<const NodeOverlap *>(&overlap) != 0
     */
-   virtual int
+   virtual size_t
    getDataStreamSize(
       const hier::BoxOverlap& overlap) const;
